refactor: user 패키지와 외부 패키지 의존성 제거 bridge 패키지를 통해서 메시지 전달

+++ src/main/java/com/poorlex/poorlex/expenditure/domain/ExpenditureRepository.java
@@ -9,25 +9,25 @@ import org.springframework.data.repository.query.Param;
 public interface ExpenditureRepository extends JpaRepository<Expenditure, Long> {
 
     @Query(value = "select coalesce(sum(e.amount.value), 0) from Expenditure e "
-        + "where e.memberId = :memberId and e.date between :start and :end")
-    int findSumExpenditureByMemberIdAndBetween(@Param(value = "memberId") final Long memberId,
-                                               @Param(value = "start") final LocalDate start,
-                                               @Param(value = "end") final LocalDate end);
+            + "where e.memberId = :memberId and e.date between :start and :end")
+    Long findSumExpenditureByMemberIdAndBetween(@Param(value = "memberId") final Long memberId,
+                                                @Param(value = "start") final LocalDate start,
+                                                @Param(value = "end") final LocalDate end);
 
     @Query(value = "select m.id as memberId, coalesce(sum(e.amount.value), 0) as totalExpenditure "
-        + "from Member m left join Expenditure e on m.id = e.memberId and e.date between :start and :end "
-        + "where m.id in :memberIds "
-        + "group by m.id ")
+            + "from Member m left join Expenditure e on m.id = e.memberId and e.date between :start and :end "
+            + "where m.id in :memberIds "
+            + "group by m.id ")
     List<TotalExpenditureAndMemberIdDto> findTotalExpendituresBetweenAndMemberIdIn(
-        @Param(value = "memberIds") final List<Long> memberIds,
-        @Param(value = "start") final LocalDate start,
-        @Param(value = "end") final LocalDate end
+            @Param(value = "memberIds") final List<Long> memberIds,
+            @Param(value = "start") final LocalDate start,
+            @Param(value = "end") final LocalDate end
     );
 
     @Query("select e from Expenditure e "
-        + "left join BattleParticipant bp on e.memberId = bp.memberId and bp.battleId = :battleId "
-        + "left join Battle b on bp.battleId = b.id "
-        + "where e.date between cast(b.duration.start as LocalDate) and cast(b.duration.end as LocalDate)")
+            + "left join BattleParticipant bp on e.memberId = bp.memberId and bp.battleId = :battleId "
+            + "left join Battle b on bp.battleId = b.id "
+            + "where e.date between cast(b.duration.start as LocalDate) and cast(b.duration.end as LocalDate)")
     List<Expenditure> findBattleExpenditureByBattleId(final Long battleId);
 
     List<Expenditure> findExpendituresByMemberIdAndDateBetween(final Long memberId,
